LayoutTests/editing/pasteboard/data-transfer-get-data-on-drop-rich-text-expected.txt

Rich text
{
    "dragover": {
        "text/html": "",
        "text/plain": ""
    },
    "drop": {
        "text/html": "<!DOCTYPE html><strong style=\"font-family: -apple-system; font-size: 150px; font-style: normal; font-variant-caps: normal; letter-spacing: normal; orphans: 2; text-align: start; text-indent: 0px; text-transform: none; white-space: nowrap; widows: 2; word-spacing: 0px; -webkit-text-stroke-width: 0px; text-decoration: none; color: purple;\">Rich text</strong>",
        "text/plain": "Rich text"
    }
}
